Individual version including character class
This file contains a complete Hurkle game using a Character Class object.  Drawing the board and scoring the game are still done via functions within the body of the program.

Suggestions: Convert Draw_Table() function to a Board Class with the Draw_Table function inside of it.  When that is functional, create a Game Class that will perform the remainder of the game functions, including instantiating and calling the other classes as needed.